refactor Image to it's own file
