+++ c_tests/test_Image.cpp
@@ -3,7 +3,7 @@
 //
 
 #include "catch.hpp"
-#include "Image.h"
+#include "TypedImage.h"
 #include "ImageFactory.h"
 #include "Reader.h"
 #include "exceptions.h"
